# api-specification.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Organix Backend API",
    "description": "API RESTful para sistema de gestão de conteúdo multi-tenant",
    "contact": { "name": "Organix Team", "email": "contact@organix.com" },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Generated server url" }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
    "/api/users/{id}": {
      "get": {
        "tags": ["User Management"],
        "summary": "Buscar usuário por ID",
        "description": "Busca um usuário específico por ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "403": {
            "description": "Acesso negado - usuários só podem ver seus próprios dados",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "put": {
        "tags": ["User Management"],
        "summary": "Atualizar usuário",
        "description": "Atualiza os dados de um usuário existente",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Dados inválidos ou email já em uso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "403": {
            "description": "Acesso negado - usuários só podem atualizar seus próprios dados",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "delete": {
        "tags": ["User Management"],
        "summary": "Excluir usuário",
        "description": "Exclui um usuário da empresa",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Usuário excluído com sucesso" },
          "403": {
            "description": "Acesso negado - apenas administradores podem excluir usuários"
          },
          "404": { "description": "Usuário não encontrado" },
          "400": {
            "description": "Não é possível excluir o administrador da empresa"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Buscar produto por ID",
        "description": "Busca um produto específico por ID",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "404": {
            "description": "Produto não encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          },
          "200": {
            "description": "Produto encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "put": {
        "tags": ["Products"],
        "summary": "Atualizar produto",
        "description": "Atualiza os dados de um produto existente",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProductRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Produto atualizado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          },
          "403": {
            "description": "Acesso negado - apenas administradores podem atualizar produtos",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Excluir produto",
        "description": "Exclui um produto da empresa",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Produto excluído com sucesso" },
          "404": { "description": "Produto não encontrado" },
          "403": {
            "description": "Acesso negado - apenas administradores podem excluir produtos"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/api/users": {
      "get": {
        "tags": ["User Management"],
        "summary": "Listar usuários",
        "description": "Lista todos os usuários da empresa com paginação",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Pageable" }
          }
        ],
        "responses": {
          "403": {
            "description": "Acesso negado - apenas administradores podem listar usuários",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/PageUserResponse" }
              }
            }
          },
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/PageUserResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "post": {
        "tags": ["User Management"],
        "summary": "Criar usuário",
        "description": "Cria um novo usuário na empresa",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Dados inválidos ou email já em uso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "403": {
            "description": "Acesso negado - apenas administradores podem criar usuários",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Listar produtos",
        "description": "Lista todos os produtos da empresa",
        "operationId": "getAllProducts",
        "responses": {
          "200": {
            "description": "Lista de produtos retornada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ProductResponse" }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "post": {
        "tags": ["Products"],
        "summary": "Criar produto",
        "description": "Cria um novo produto na empresa",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProductRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Dados inválidos",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          },
          "403": {
            "description": "Acesso negado - apenas administradores podem criar produtos",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          },
          "201": {
            "description": "Produto criado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateUserRequest": {
        "required": ["adminType", "email", "name"],
        "type": "object",
        "properties": {
          "name": { "maxLength": 100, "minLength": 2, "type": "string" },
          "email": { "type": "string" },
          "password": { "maxLength": 100, "minLength": 6, "type": "string" },
          "adminType": { "type": "string", "enum": ["ADMIN", "OPERATOR"] }
        },
        "description": "Novos dados do usuário"
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "adminType": { "type": "string", "enum": ["ADMIN", "OPERATOR"] },
          "companyId": { "type": "string", "format": "uuid" },
          "companyName": { "type": "string" }
        }
      },
      "UpdateProductRequest": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": { "maxLength": 100, "minLength": 0, "type": "string" },
          "description": { "maxLength": 500, "minLength": 0, "type": "string" },
          "category": { "maxLength": 50, "minLength": 0, "type": "string" }
        },
        "description": "Novos dados do produto"
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "category": { "type": "string" },
          "companyId": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateUserRequest": {
        "required": ["adminType", "email", "name", "password"],
        "type": "object",
        "properties": {
          "name": { "maxLength": 100, "minLength": 2, "type": "string" },
          "email": { "type": "string" },
          "password": { "maxLength": 100, "minLength": 6, "type": "string" },
          "adminType": { "type": "string", "enum": ["ADMIN", "OPERATOR"] }
        },
        "description": "Dados do novo usuário"
      },
      "CreateProductRequest": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": { "maxLength": 100, "minLength": 0, "type": "string" },
          "description": { "maxLength": 500, "minLength": 0, "type": "string" },
          "category": { "maxLength": 50, "minLength": 0, "type": "string" }
        },
        "description": "Dados do novo produto"
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": { "minimum": 0, "type": "integer", "format": "int32" },
          "size": { "minimum": 1, "type": "integer", "format": "int32" },
          "sort": { "type": "array", "items": { "type": "string" } }
        }
      },
      "PageUserResponse": {
        "type": "object",
        "properties": {
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserResponse" }
          },
          "number": { "type": "integer", "format": "int32" },
          "sort": { "$ref": "#/components/schemas/SortObject" },
          "pageable": { "$ref": "#/components/schemas/PageableObject" },
          "numberOfElements": { "type": "integer", "format": "int32" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" },
          "empty": { "type": "boolean" }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": { "type": "integer", "format": "int64" },
          "sort": { "$ref": "#/components/schemas/SortObject" },
          "pageNumber": { "type": "integer", "format": "int32" },
          "pageSize": { "type": "integer", "format": "int32" },
          "unpaged": { "type": "boolean" },
          "paged": { "type": "boolean" }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": { "type": "boolean" },
          "sorted": { "type": "boolean" },
          "unsorted": { "type": "boolean" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Token para autenticação. Formato: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}