# src/main/java/com/organixui/organixbackend/product/dto/ProductResponse.java
package com.organixui.organixbackend.product.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.UUID;

/**
 * DTO para respostas contendo informações de produtos.
 */
@Data
@NoArgsConstructor
@AllArgsConstructor
public class ProductResponse {
    private UUID id;
    private String name;
    private String description;
    private UUID companyId;
    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;
    private String createdAtFormatted; // Data de criação no formato dd/MM/yyyy
}